operator/tests/pyproject.toml
[project]
name = "operator-tests"
version = "0.1.0"
description = "Operator tests"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
    "sh>=2.2.2",
]
